feat: Implement placeholder functions and integrate NLP into brain
Implemented two key placeholder functions:

1. **Salience Evaluator**: Fixed hardcoded num_features placeholder
   - Added brain_get_num_inputs() getter function to brain API
   - Updated salience_evaluator_create() to use actual brain config
   - File: src/cognitive/salience/nimcp_salience.c:681

2. **Neural Network Neuromodulation**: Implemented dopamine level query
   - Added neuromodulators header include
   - Implemented neural_network_get_neuromodulation() to query dopamine
   - Returns actual NEUROMOD_DOPAMINE level instead of 0.0f placeholder
   - File: src/core/neuralnet/nimcp_neuralnet.c:2846

Additional changes from previous session:
- Integrated NLP network into brain structure
- Added emotional system stubs
- Reorganized test structure
- Various build system updates

Changes:
- src/core/brain/nimcp_brain.h: Added brain_get_num_inputs() declaration
- src/core/brain/nimcp_brain.c: Implemented brain_get_num_inputs() and NLP integration
- src/cognitive/salience/nimcp_salience.c: Use brain_get_num_inputs()
- src/core/neuralnet/nimcp_neuralnet.c: Implement neuromodulation query

+++ src/security/nimcp_security.c
@@ -1453,7 +1453,7 @@ nimcp_result_t nimcp_encryption_generate_key(uint8_t* key)
     srand(seed);
 
     for (int i = 0; i < NIMCP_SECURITY_KEY_SIZE; i++) {
-        key[i] = (uint8_t) ((rand() ^ (rand() << 8)) % 256);
+        key[i] = (uint8_t) ((rand() ^ ((unsigned int)rand() << 8)) % 256);
     }
 
     return NIMCP_SUCCESS;
